Compressed all images, reorganized file tree
Images were way too large and taking way too long to load on site, so I compressed them all. The file tree was also a mess, so I reorganized that to actually make sense and to also be clean. I also added a maintenance.html file for when the site is under going construction